pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ambient"
version = "0.1.0"
description = "Sleep biometrics monitoring using TI IWR6843AOPEVM mmWave radar"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
	"Development Status :: 3 - Alpha",
	"License :: OSI Approved :: MIT License",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
]
dependencies = [
	"numpy>=1.24.0",
	"scipy>=1.10.0",
	"pyserial>=3.5",
	"structlog>=23.1.0",
	"pydantic>=2.0.0",
	"h5py>=3.8.0",
	"pyarrow>=12.0.0",
	"pandas>=2.0.0",
	"click>=8.1.0",
	"rich>=13.0.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.7.0"]
dev = [
	"pytest>=7.3.0",
	"pytest-cov>=4.1.0",
	"ruff>=0.1.0",
	"black>=23.0.0",
	"mypy>=1.4.0",
	"pre-commit>=3.3.0",
]

[project.scripts]
ambient = "ambient.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120
target-version = ["py310"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
	"hardware: requires physical radar",
	"slow: slow test",
]
